LayoutTests/fast/forms/number/number-skip-spaces-in-user-input-expected.txt

should skip spaces from user input number

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.valueAsNumber is 123456
PASS input.value is "123456"
PASS successfullyParsed is true

TEST COMPLETE
